requirements-freeze.txt
absl-py==2.1.0
affine==2.4.0
aiobotocore==2.12.3
aiohttp==3.9.5
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
bentoml==1.2.12
billiard==4.2.0
boto3==1.34.69
botocore==1.34.69
build==1.2.1
cattrs==23.1.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circus==0.18.0
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
configobj==5.0.8
contextlib2==21.6.0
cryptography==42.0.5
deepmerge==1.1.1
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.49.0
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.1.0
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
filelock==3.13.4
fiona==1.9.6
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.3.1
funcy==2.0
geopandas==0.14.3
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
grpcio==1.62.2
gto==1.7.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
importlib-metadata==6.11.0
inflection==0.5.1
iterative-telemetry==0.0.8
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.7
lightning==2.2.2
lightning-utilities==0.11.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==11.525.150
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==3.11.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydot==2.0.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyparsing==3.1.2
pyproj==3.6.1
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
rasterio==1.3.10
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2024.3.1
s3transfer==0.10.1
schema==0.7.5
scmrepo==3.3.1
semver==3.0.2
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simple-di==0.1.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
sqltrie==0.11.0
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tomli_w==1.0.0
tomlkit==0.12.4
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
voluptuous==0.14.2
watchfiles==0.21.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1
